AFS-60 Why Are we not Tracking Clicks for Tremor and APNXs?
It looks like our VAST click handler is not firing an event to
programmatic tracking service. We seem to be triggering 3rd party
embedded trackers in the content but not our own for tracker.

The fix is to send a tracking event for click in the overlay event
handler when the user clicks.