ABT-333 iOS Blackscreen on specific SDK versions
The first issue is that on a cached file request we don't check for
result code of 0 on iOS.

The next issue is that DOMParser's parseFromString seems to have some
kind of race condition that causes the documentElement to be null
occasionally.

The fix is to always just use the polyfill for that method on iOS.
Added DOMUtils and moved that method into there as a utility.
Added unit tests for utility method.