Extracted campaign image logic to a function.
Also fixed an issue where same asset got cached twice (the custom image). This happened because AssetManager was using asset.getUrl() which returned the already cached url that had 'file://' in it and since that url with the 'file://' wasn't the originally cached url it got cached again causing the url to have double 'file://'.