Ignore areControlsVisible in showCallButton.
This check stopped the call button from coming visible if the other controls had were not visible. We already made a change last year where the call to action button stays visible even when other controls are faded out so this check is unnecessary. Before this fix if user fades the controls away before the call to action button is shown, the call to action button won't be shown either until user manually fades the controls back in.