Add temporary custom Kafka event to performance video overlay download.
This ads custom Kafka event from the Performance video overlay install now button with information when the button was pressed and what was rating of the game. We already had similar event before but reason adding this back is that we have new version of the button that shows rating of the game. We want to know is there correlation with the game's rating and how likely user will press the button.