Fix iOS crash when orientation doees not match supported.
The safety check usees information from the project's
general -> Deployment Info -> Device Orientation
to determine if a orientation is supported.  If it isn't the
orientation is set to one of the supported orientations.

This causes a problem when a developer manually changes the
orientation in code.  In this case, the orienttation is correct
and gets set incorrectly as that orienatation is not in
supported orientations.  Forcing the orientation to one in
the supported orientations causes a crash.